Add Gitlab changelog configuration file
This commit adds the configuration file for the Gitlab changelog
generation [1]. The configuration file has to be located at
".gitlab/changelog_config.yml". Currently (so in this branch/mr) it is
not respected. However, my guess is that this will be the case after
we merge it into the default branch of the repository (so "devel").

The configuration file was inspired by [2].

[1] https://docs.gitlab.com/ee/api/repositories.html#generate-changelog-data
[2] https://gitlab.com/gitlab-org/gitlab/-/blob/df07064699884ac3881ea3b74ecb95bf36dfedac/.gitlab/changelog_config.yml

Changelog: added

f