haproxy: allow proxy-protocol for individual ports
"The PROXY protocol provides a convenient way to safely transport
connection information such as a client's address across multiple
layers of NAT or TCP proxies." [0]

This commit adds an option to toggle the use of the proxy protocol for
indivual lb-ports. It's a handy feature for situations where HAProxy
acts as a L4-LB, i.e., it simply forwards traffic, and the backend needs
certain information about the connection, e.g. the source IP. This
information would otherwise be lost because of NAT. `X-Forwarded-For`
isn't suitable for these situations because TLS is not terminated in
HAProxy but in the backend so no additional information can be added to
the opaque data stream.

[0] http://www.haproxy.org/download/1.8/doc/proxy-protocol.txt