Fix: legacy converter: ceph fs pool replication variable
The variable has been renamed to "(rook_)ceph_fs_replicated" in [1][2][3].
We are using a default of "1" in the rook role (default variable file) [3].
This means that the pool size will be decreased to "1" for clusters which
are using a config that has been created by the legacy converter (sic!).

This commit fixes the legacy config converter such that the variable
has the correct key and is ensured to be respected.

[1] 74666e85879790651ece45d7edf53f7e32e2743e
[2] 45d8c32b34d8a69f98c8d7047c04e9bf48cc4a66
[3] https://gitlab.com/yaook/k8s/-/blob/022b2d1a581055a8f760621b233ab4fdca0ba602/k8s-service-layer/roles/rook_v1/defaults/main.yaml#L42

Note that the respective MR which introduced the renaming is a tombstone.
It has been renamed when adding support for multiple rook versions and
refactoring the inventory updated / configuration file layout. The new
inventory handling automatically applies the "rook_" prefix to the
variables of the rook section (in the configuration file). Because
e.g. "fs", "fs_name", "fs_replicated" and "fs_preserve_pools_on_delete"
did not seem very meaningful (especially "fs"), the "ceph" prefix has
been added.