Fix: k8s upgrade: Ensure kube-proxy is removed
This commit ensures that kube-proxy is removed during a k8s upgrade,
if kube-router is used as CNI.

Running "kubeadm upgrade apply" recreates the kube-proxy DaemonSet [1].
It seems like it is currently not possible to avoid that [2]. This is a
problem, because kube-proxy and kube-router are in conflict. It is
important to remove the kube-proxy DaemonSet again AND cleanup the
iptables during a k8s upgrade. Otherwise k8s service endpoints won't be
reachable.

Removing the kube-proxy ConfigMap prior to the upgrade does not help as
well (like it is suggested in [3]), because kube-router needs access to
it and if the ConfigMap is removed, kube-router Pods will be "Completed"
instead of "Running".

In addition this commit makes the tasks conform to the style guide [0].

[0] https://gitlab.com/yaook/meta/-/wikis/Development-Workflow
[1] https://kubernetes.io/docs/tasks/administer-cluster/kubeadm/kubeadm-upgrade/#how-it-works
[2] https://github.com/kubernetes/kubeadm/issues/1756
[3] https://docs.cilium.io/en/v1.9/gettingstarted/kubeproxy-free/