Rework how nodes get their taints and labels
The previous approach with a role in stage 4 had multiple downsides:

- The taints and labels could not be relied upon in stage 3.
- It requires admin access to the k8s cluster (i.e. it cannot be used to
  autonomously spawn a node).
- It may modify existing nodes, which can cause disruption (e.g. if a
  NoExecute taint is accidentally added to a crucial storage node).
- It used string interpolation, which opened the door for abuse for
  instance to remove labels by suffixing with `-`.
- Taints only take effect late, which means that workload which must not
  (ever) run on a node (e.g. for GDPR reasons) may briefly be scheduled
  there.

As disruption caused by taints+labels cannot easily be reasoned about,
the only way out keeping this would be to make the task always
disruption-requiring. That renders it effectively moot.

The new approach uses init/join configuration for kubeadm in order to
apply labels and taints right during node initialization. This allows
nodes to join the cluster with their taints and labels configured
correctly right away. As this only affects the initial join, no
disruption can be caused by this. It also works without extended
privileges.

A crucial downside of this approach is that the set of taints and labels
which can be set may be restricted by Kubernetes. As we generally do not
set those, it doesn't affect us at this point.