Inventory: Move ssh-hardening and etc vars to vars_files
This commit moves the ssh (hardening) variables to a newly created vars
file "ssh-hardening.yaml". The etc variables (path to known hosts and
etc directory as well as the ssh extra arguments) have been moved to
a new vars file "etc.yaml" as well.

The ssh-hardening vars are needed in stage2, stage3 and for upgrades. We
are now able to include them if needed. Same applies to the etc vars.

Issue: #346