Move k8s API server gateway preparation to stage3
Similar to the haproxy/nftables refactoring where the gateway preparation
has been moved to the "global-monitoring-support" role which is executed
in stage3, this commit does the same for the Kubernetes API server.

This has the advantage that we k8s api server frontend and backend port
are configurable now and no workaround to move them to stage2 is needed
anymore. Related MR: !322

Issue: #346