SLS: Make monitoring checks independent
The monitoring checks have been executed from the first control plane
node. That had several downsides, first a connection to that node is
needed, but the requirement we stated for the KMS layer is, that the
only thing you need is the kubeconfig. Because a connection to the
first control plane nodes was needed, we had to check connectivity to
that node (detect_user role) which required us to pass over (at least)
the "etc_dir" variable, because it is needed by the detect_user role.

This commit adjusts the monitoring checks such the scraping happens
from within a static Pod.

---

SLS ... Service Layer Separation