Config: Add comment about storage class interdependency
The rook role expects to find the "local-storage" storage class.
However, !383 made these names configurable and more important, disabled
them by default. This can lead to a failure of the rook role if rook
is enabled, but there is no storage class.

This commit adds a comment and the rook_mon_volume_storage_class to the
configuration template.

Changelog: added