tests: Add default values for local storage classes
!383 made the local storage classes configurable. However, providing
a storage class name is optional. If it is not defined inside the
config.toml, it won't be available to the check-local-storage checks
and the smoke tests fail. The setup itself won't fail, because the roles
have the storage class name defined in "defaults/main.yaml". Outside of
the role the name won't be available.

The tests should be decluttered in the near future anyway so I am going
for the quick solution and added in-line defaults.

Changelog: fixed