Wireguard: make C&H users default to false
We agreed on defaulting the C&H related configurations to false.

Additionally, this commit removes a unnecessary check if the submodule
directory does exist. That function calls is now guarded. If the user
configured to rollout/use the C&H specific wireguard users repository,
the inventory updater should fail if the directory does not exist.

Issue: #324

Changelog: changed