Wireguard: Fix configuration generation w/o C&H users
The "wireguard_company_users" have been initialized as a list. However,
when calling the "_merge_wg_users" we do pass the ".values()" of
wireguard_company_users (which is not possible for a list).

This error only happens if C&H users are disabled, because otherwise
the "_load_wireguard_company_users" would ensure that
"wireguard_company_users" is a dictionary.

This commit ensures that "wireguard_company_users" is initialized as
a dictionary and additionally ensures that at least one wireguard peer
is configured before calling "_require_unique_names" (because there
we are using ".values()" again). Enabling wireguard but not configuring
a peer makes no sense and should fail anyway.

Issue: #324

Changelog: fixed