hosts-file: Fix boolean declaration of ini inventory
Only "True" and "False" are interpreted as boolean values by ansible.
Everything else is considered a string!

When checking values of such "boolean" variables without a "| bool"
filter, "false" would be considered true because it is a non-empty
string. This will cause havoc!

This has been noticed when trying to deploy a baremetal cluster with
"dualstack_support=false" in the ini inventory. This variable is
checked at ~50 different locations and mostly without the boolean filter
(" | bool").

We did not notice that yet because "on_openstack" is checked only
once, and is set to true (in e.g. CI) so the non-empty string applies,
has a default (so if you do not define it, a proper bool value is used)
and is converted to bool by the boolean filter.

This commit adjusts the value to have capital starting letter so that
developers do not get confused how to write a boolean value.

https://docs.ansible.com/ansible/latest/collections/ansible/builtin/ini_inventory.html