checks: Reenable check-services with lb ports
tl;dr: The checks could never be triggered

The checks could never be trigger, because "deprecated_nodeport_lb_test_port"
is a gateway specific variable of stage2. These variables were not
available to the tests playbook and therefore the check that
"deprecated_nodeport_lb_test_port" is defined was always false.

In addition, the frontend "gateway" variable determination was incorrect.
If the group "gateways" exists, the whole group has been chosen.

On top of that, the check-services templates had hardcoded ports.

The tests are currently a mess and should be decluttered.

Issue: #361