frontend-prep: Configure nftables only if on OpenStack
The frontend preparation has been partially moved to stage3 in
yaook/k8s!392 and during the inventory refactoring (tombstone). These
roles are executed against gateways only, so if there are no gateways,
e.g. if you are on bare-metal, these roles are not triggered and the
configuration does not happen. This is bad, because in that case you
cannot access your k8s-API and global monitoring is not working.

This commit changes the hosts for these roles to "frontend" so that it
is ensured they get executed. This required to make the nftables
configuration optional, because in the bare-metal case there is no
nftables running on the frontend nodes (control-plane nodes) and the
roles fail. Basically, we want to differentiate if we are running on
OpenStack or on bare-metal, so the "on_openstack" variable has been
chosen as guard for the nftables configuration tasks.

In addition this commit introduces a handler to restart nftables (#336).