Add possibility to deploy Vault instance in Docker
In order to prepare the migration from local stored
certificate and secrets to a centralized Hashicorp Vault
instance, we need an easy way to deploy a Vault instance
with Docker.

Created bash scripts to that will set up Vault in Docker and
TLS enabled.

`vault.sh` will take care of the container creation and will trigger
`vault_ssl.sh` to create the needed certificates.
The script will check if a Vault instance is already
there and based on that either create one or start it again.
It's designed to be reused, in case container went down.
It checks init and seal status before doing so.

'vault_ssl.sh' will create TLS root and host certificates incl.
secrets and chain them together for the Vault HTTPS server.

'vault_env.sh' can be used to source the Vault env vars.
It also provides infos to the user about the used Vault client and
server versions, the URL incl. the port to the Vault instance, and
the for a login needed token and CAcert infos.

'envrc.template.sh' got a switch to activate the usage and creation
of the Vault instance, a possibility to specify where all Vault related
data will be stored, and a possibility to source the 'vault_env.sh'.

'config.template.hcl'
Vault config file template, that will be copied during
vault.sh execution. TLS and UI have been enabled by default.

'.gitlab-ci.yml'
The test is designed to test our 'config.template.hcl'
default settings. The certificates and config are stored within
the CI vars, as the service instance is limited in accessing
data from the repo while it's starting.