CI: Create manual jobs only for MR events
Currently we're seeing pipelines with manual cluster jobs which have not
been triggered being marked with an exclamation mark ("!"). However, the
manual jobs did not get triggered. If you hover over the job in the
pipeline overview, you see that the job is in state "forward deployment
failure". This may be because we're using a protected environment for
the cluster-spawn jobs.

However, creating the manual jobs for non-MR related events does not
really make sense anyway so this commit adjusts the CI manifest such
that these jobs do get created only there.