monitoring: Use helm chart to install prometheus stack
This commits creates new roles (monitoring_v2) in the KSL and KMS layer
which do deploy the monitoring stack using helm charts.

The prometheus stack is now deployed via the community helm chart [1].
The main reason is to simplify the implementation. The previous
implementation [2] (monitoring_v1) used JSONNET to create the monitoring
stack. JSONNET has a steep learning curve and only selected individuals
were able to fully understand the previous setup.

The prometheus operator is deployed in advance to the actual prometheus
stack because we observed timeouts if they are deployed in parallel.
A lot of components are enabled by default by the helm chart so I needed
to take care that things will not be duplicated.

We're leaving out monitoring of etcd for now as it requires
authentication [3].

This commit (and mr) has left thanos out of the helmification for now.
This decision is based on the desire to get a working monitoring stack
via helm more quickly.

Issue: #375 #366

[1] https://github.com/prometheus-community/helm-charts/tree/main/charts/kube-prometheus-stack
[2] https://github.com/prometheus-operator/kube-prometheus
[3] https://github.com/helm/charts/issues/6921