monitoring: Make helm chart version configurable
To leave the delta of changes as minimal as possible for existing
clusters, we're implementing support for the helm chart version that is
as closest to the previous JSONNET implementation.

As a follow up step, we'll need to implement the corresponding upgrade
path(s) to the latest version (upgrades require additional steps [1]).

Issue: #375 #366

[1] https://github.com/prometheus-community/helm-charts/tree/main/charts/kube-prometheus-stack#upgrading-chart