monitoring: Integrate thanos into prometheus helm stack
This MR does not address the implementation of the option to deploy
thanos via a helm chart (like the prometheus stack). Instead, we're
cutting out thanos from the JSONNET implementation and integrate it
into the monitoring_v2 setup. The main reason is to get a working
monitoring stack deployed via helm more quickly.

This commit cuts out thanos from the JSONNET implementation and adds
thanos support to the prometheus stack helm implementation.

Issue: #375 #366