monitoring: Enable monitoring of kube-proxy
By default kube-proxy will provide metrics on 127.0.0.1. This is to
restrictive and Prometheus fails to gather metrics (connection refused).
The metrics have to be exposed on 0.0.0.0. The port has to be specified.
10249 is the default port. I did not see any use on making it configurable
(for now).

Note that we'll need to adjust the ConfigMap for kube-proxy in existing
clusters (if we do not do this, scraping fails).

We currently do support two CNIs: kube-router and calico. [*]
kube-router and kube-proxy are in conflict. Side fact: this fell on our
feets during kubernetes upgrades (rel !385).

kube-proxy is only present if calico is used as CNI so this commit
ensures that kube-proxy is only monitored if that's the case.

Issue: #375 #366

[*] Basically, while writing this commit we do also state support for
    flannel, but we're planning to get rid of it in the near future.
    More infos in the respective issue (to be created).