gpu-support: CentOS is dead. Long live Ubuntu!
This commit takes the gpu-support role apart and makes it usable for
Ubuntu 18.04 hosts. At the same time support for CentOS was removed for
two reasons:

- CentOS itself is somewhat EOL
- the RHEL-* part of this role cannot be maintained due to time
  constraints

If you have a CentOS-based k8s worker, your best course of action is to
get rid of it and re-create from scratch with the Ubuntu image you're
already using.

Furthermore we're no longer using the runfile installation but make use
of nvidia's package repository.