docker: clarify template creation & service restart
Under special circumstances it was possible that an incomplete docker
configuration (daemon.json) was deployed on a node. Also a confusing
number of notifications were spread across the role. Handlers, normally
a good thing, don't cut it here because of the required, additional
logic.

This commit clarifies the logic and gets rid of these handlers.