ci: Replace an 'L' node with a GPU-enabled one
The GPU code path of our LCM can and will from now on be regularly
tested. managed-k8s-worker-5, the node with the changed flavor, hasn't
been assigned any special storage or monitoring activity so it should be
free to host a tiny deviceplugin daemonset and cuda job.

See https://redmine.cloudandheat.com/issues/33233 (C&H-only) for
details on the flavor.