lbaas-test: use random port for curl check
ch-k8s-lbaas is pretty clever in a sense that it will re-use an already
allocated floating IP address if no other service shares the same
port. In our test suite we use port 80 which is a common port. That
means if a customer uses LoadBalancer-typed services we cannot use that
floating ip because port 80 is already "occupied". This in turn means
that the project quota needs N+1 floating IP s.t we can run our tests
which is unfortunate. By using a less common port we reduce the "risk".

Context for C&H folks: https://redmine.cloudandheat.com/issues/33352