Move k8s base plays to dedicated directory
Following the style of the kms and ksl layer. Making structure consistent.

This commit also renames the stage2 and stage3 playbook to be compliant
to the naming scheme of the upper layers. Same applies to the the tests.

The "ansible" directory now contains only ansible related files. Nameley,
the ansible configuration (which is now set as env var in lib.sh) and
the galaxy requirements.

The helper plays have been moved to an own directory as they're used
rather irregularly. Same applies to the k8s upgrade plays. Paths inside
the playbooks needed to be adjusted due to that.

Issue: #346