CI: Create job which tests k8s upgrade path
Obviously more than overdue. This commit introduces a new CI stage which
consists of a job which checks the upgrade path from the minimal
supported k8s version to the latest supported one. The upgrade CI job
manifest is basically an extension to the spawn-cluster job manifest,
and therefore makes heavy use of [1].

The upgrade path is tested for both CNI plugins we currently do support
(namely calico and kube-router).
Despite the CNI plugin should not make any difference in theory, it does
in practice, as kube-router is in conflict with kube-proxy which gets
recreated by kubeadm during upgrades (refer !385).

Testing the k8s upgrades with calico as CNI plugin makes the job logic
a bit more complicated due to the following circumstances:
AFLAGS get overwritten by the script procedure (because we need
to adjust it during upgrades as we do not want to tinker around
with the CI config out of the shell). We need to make the
k8s_network_plugin variable persistent throughout the whole job.
To realize that, the AFLAGS_BASE expansion is used.

Additionally this commit deduplicates variable definitions.

Issue: #393

[1] https://docs.gitlab.com/ee/ci/yaml/yaml_optimization.html#reference-tags