monitoring: Extend jsonnet-based monitoring support
Make the jsonnet-based monitoring stack configurable. We do need this
for existing productive clusters which may don't like it when we
completely tear down their monitoring to build it again via helm.
Therefore, we do make the monitoring_v1 role configurable and give them
more time to prepare mentally for it.

The helm-based monitoring setup is configured by default, so be careful.

After all existing clusters have been migrated to monitoring_v2, we
should purge the jsonnet-based monitoring.

As we had a change of perspective regarding the Service-Layer-Separation
and the monitoring stack setup in !415 [*], this commit moves the thanos
role to the KSL.

Issue: #375 #366

[*] The original view was that the KSL should built the "basic" for the
    monitoring stack, but no components itself. The new point of view is,
    that the KSL builds the monitoring stack and ensures that all base
    kubernetes components are monitored. Architecture / setup specific
    (e.g. monitoring the ch-k8s-lbaas) should be setup by the KMS layer.