Add action for system updates of initialized nodes
Installing system updates is generally a disruptive task because it's
hard to reason about what's going to happen. Is the docker runtime
restarted? Systemd doing some fancy things to the network?
`update_system` is therefore only allowed if the node isn't used yet
(evaluated by a heuristic) or if `MANAGED_K8S_RELEASE_THE_KRAKEN=true`
is set. The latter also allows the playbook to do other scary things so
I hardly invoke stage-2 and stage-3 with no holds barred.

"But you could limit execution to a subset of tasks!" Sure, I could. But
it's easy to forget the tag and the current implementation actually harms
performance. On the set of existing nodes, the package cache is updated
*in serial* which takes a long time with a sufficient number of nodes.

In order to resolve the above, this commit

- does no longer update nodes anyhow that are already initialized (both
  gateways and k8s nodes)
- adds another action `system_update_nodes.sh` if, in case you're
  using the LCM for the updates