calico;kube-router: Implement CNI migration tasks
kube-router is having scaling issues. We're experiencing high CPU usage
for rather small clusters. Also supporting two CNI plugins is complex
bloats out test matrix. In a secret council of wise men on a cloudless
night of a full moon it has been decided that the support for kube-router
will be dropped and the live migration needs to be implemented.

This commit enables migrating a yk8s cluster from kube-router to calico.

The necessary steps have been put into an own role because this solution
is minimal invasive to the existing code base and can easily be removed
after all existing clusters have been migrated.

NOTE that these tasks are execute separately on workers and masters.

We need to remove all kube-router instances (or at least the DaemonSet)
and clean up the drop-in network configuration directory on all host
nodes. As kube-router and kube-proxy are in conflict, but calico and
kube-proxy are not, we do (re-)install kube-proxy. The simplest way to do
that is to use kubeadm (please do not reveal that this is possible,
because the documentary is silent about it). Changing the CNI plugin
also requires to restart nearly everything. The simplest way to achieve
that is to use kubectl to rollout restart in all namespaces.
Obviously this is disruptive and should be executed only if necessary.

Issue: #402