monitoring_v1: Fix file lookup error handling
This commit fixes the condition for which the error handling is set to
strict. This is necessary for fresh clusters and the CI as otherwise the
lookup will fail (if the default configuration is used and monitoring_v2
instead of monitoring_v1 is rolled out) because the file
"jsonnet-sources/manifests.json.files" is created dynamically during roll
outs and is not present if monitoring_v1 has never been executed.

Issue: #375 #366