git: Do not git ignore hosts files of stage2&3
These are needed throughout the plays. Operation of a cluster is very
restricted if they are not available.

These files are generated by terraform. Therefore, to generate them, one
needs to source the appropriate OpenStack openRC file (and have the
credentials in place). You do not want to have to remember this at night.