Upgrade supported Ansible version to 5.3.0
This commit upgrades the supported Ansible version to v5.3.0.
As the upgrade / migration path is not trivial and not in the scope of
this LCM, it is left out. As users should use virtual environments,
it should be super easy for everyone to setup a new one and skip the
upgrade migration.

The upgrade to a newer Ansible version is necessary (and a bit overdue)
as the cryptography module throws an exception with the current setup
(refer #388).

Closes: #388