Update Kubernetes and openshift packages
Pinning the Kubernetes package is not necessary anymore [1]. In fact,
it also breaks our setup after upgrading Ansible.

We need to set state to "latest", otherwise the pip packages won't get
updated. Another option may be to create a separate task for each
pip package we want to install and then use the "version" flag [2].
I did not test if this updates a package. Also we may make the "state"
flag dependent on "_allow_disruptions" or find time to define a proper
update flag/mechanism. Splitting "k8s-tools" out like in !429 may also
be worth a try. If somebody is reading this and has spare time, go ahead.

[1] https://github.com/kubernetes-client/python/issues/1333#issuecomment-992369110
[2] https://docs.ansible.com/ansible/latest/collections/ansible/builtin/pip_module.html#parameter-version