Adjust handler execution
The `Restart kubelet` handler isn't triggered, because handlers
get triggered only when a task has completed with the a `changed`
status.
This resulted into, that the kubelet service wasn't enabled and
k8s didn't came up after a restart.

Moved the handler notification to `Install kubelet, kubeadm and kubectl`
tasks.