Make Calico ipipmode configurable
User changed ipipmode settings on the Calico ippool's, are always
overwritten with every subsequent stage 3 execution.

This applies e.g. for deployments through the MetalController.

Added a ipipmode key into the config.toml template, which if activated,
will set the corresponding value during the application of the
ippools.yaml template.