monitoring: Bundle prometheus stack installation
On early adaption attempts of the kube-prometheus-stack, we experienced
timeouts during installation. It was suggested to split the operator
installation from the rest. However, this may do more harm as the
operator cannot be found under the same release name as the rest of the
stack.

We are using a way more recent stack version now as back then and the
timeout problem has not re-occurred in my tests. Therefore, this commit
simplifies the kube-prometheus-stack installation by bundling it again.