Allow spawning nodes without any taints
This allows to remove the default "master" taint right during bootstrap,
if desired. Due to the refactor of the labelleing/taint logic in
4baba5e9, the old hack of appending a `-` to the taint name does not
work anymore. Hence, there is now no way to use the LCM to spawn master
nodes without any taints.

This should fix that.

This addresses the root cause of a problem found by @mauhau in [1].

   [1]: https://gitlab.com/yaook/incubator/installation-guide/-/merge_requests/5#note_800811833