CI: Hotfix monitoring stack installation
In the ci/config/config.toml we set
k8s-service-layer.prometheus.thanos_objectstorage_container_name to
"managed-k8s-monitoring-thanos-data" which is wrong. The object store
container name CAN NOT be chosen freely. It is actually a concatenation
of the terraform.cluster_name and "--monitoring-thanos-data" [1].

As we set terraform.cluster_name to "ci" in the ci/config/config.toml,
we MUST set k8s-service-layer.prometheus.thanos_objectstorage_container_name
to "ci-monitoring-thanos-data".

The symptoms of mismatched container names were that thanos fell into
CrashLoopBackOff and prometheus did not come up which we noticed
in the monitoring_v2_test.

https://gitlab.com/yaook/k8s/-/blob/2a9634d4c2e1e95d5af7bdc6a1ab426fec3e7997/terraform/50-object-storage.tf#L2[1]