CI: Ensure OS project is clean in before_script
This commit adds a script which removes all (yaook/k8s relevant)
resources from the currently sourced OpenStack project.

The script is added to the before_script steps for spawn_cluster jobs.
We do need to ensure that the OpenStack project is clean as the quota
exactly fits the requirements of a CI cluster. If a job gets canceled,
the cleanup.sh script does not get executed and the project is filled up
with zombie resources. There are situations where this happens without
manual interaction (cancellation):

We do use the merge train feature. The merge train auto-cancels redundant
pipelines [1] which is a problem as we currently are able to run two jobs
in parallel. This means that if one fails, it can cancel the other job.

Imagine the following merge train queue: MR1-MR2-MR3.
We do have two runs in parallel, one is executing a job in MR1 and the
other in MR2. Now the MR1 job fails. As MR2 is based upon MR1 and MR1 now
cannot be merged as the pipeline failed, the job of MR2 is redundant and
gets canceled. The new merge train queue will look like: MR2-MR3 but
these merge requests cannot be merged as half of the jobs will fail as
there are zombie resources left in one of the projects.

[1] https://docs.gitlab.com/ee/ci/pipelines/merge_trains.html#automatic-pipeline-cancellation

Really helpful were:
[2] https://github.com/openstack/openstacksdk/tree/master/examples
[3] https://opendev.org/openstack/ansible-collections-openstack/src/branch/master/plugins/modules