Make network MTU configurable in terraform
This is useful for our new clouds, where we can offer MTU 1500 (and
beyond, if you dare) for tenant networks, which provides a better user
experience.

Note that we `lifecycle.ignore_changes` the MTU, because at least with
OpenStack train, a changed MTU on a port works fine initially, but then
breaks live migration (something about mismatching port MTUs on source
and destination). Potentially this could be fixed by messing with the
database, but factually this is an OpenStack bug (it should either work
or raise an error, not fail in non-obvious ways later on) which we need
to work around.