grafana: load admin creds from secret
Grafana is stateless by default which means that if a Grafana pod dies
any changes made to the running instance are lost, including updated
admin passwords. This commit externalizes the admin credentials into a
separate Secret. It's created by default but will not be overwritten by
the LCM after a user change.