check-ipsec: Add retries
This role may fail due to normal network hickups which is annoying as
this then results in the failure of a CI job very late. As a consequence,
the merge train crashes (visually spoken) and one have to retry and wait
another hour for the spawn cluster jobs to succeed.

This commit adjusts the check-ipsec role such that the k8s_exec Ansible
module is used to ping the test-VM and that ping is retried to make sure
the IPsec tunnel is not established and not working for real.