tf: Create thanos' object container on demand
To put it simple: if we're not using thanos, terraform does not have to
create an object storage container for it. Because of the way we're
currently handling variables, this commit introduces a tiny hack to
update_inventory which reads the `use_thanos` variable from the KSL
section and writes it into the terraform section. In the long run, we
probably don't want these strict scopes because our layer separation is
not as watertight as we hoped.