Fix firing alert KubeletDown by adding relabeling
The alert KubeletDown fires because it expects a scrape job
with the labels `job=kubelet` and `metrics_path="/metrics"`.
The second is missing because there is no relabeling config
which adds the `metrics_path` label.

The kube-prometheus-stack helmChart sets this
but we are overriding the relabelings.
https://github.com/prometheus-community/helm-charts/blob/kube-prometheus-stack-10.3.5/charts/kube-prometheus-stack/values.yaml#L773-L774

Set the relabeling again.