config: Increase defaults for rook operator resources
The default limits of 0.1 CPU cores and 256 MiB RAM are not sufficient
for any non-trivial cluster. Therefore, this commit increases these to
1 CPU and 512MiB RAM which should be enough for most use cases.

If a setup needs further resources, it has reached a state of
peculiarity where setting the limits via the config.toml becomes a
reasonable burden for the operator.

Closes: #451