Forbid two LBaaS controllers running at the same time
The Kubernetes Deployment will, by default, allow two pods to co-exist
(the default for maxSurge is 25%, and the absolute number of pod
overshoot is calculated by multiplying that number with the number of
replicas and then *rounding up*). This would be potentially fatal for
ch-k8s-lbaas-controller, so let's avoid that.