is-k8s-installed: Use more sophisticated checks
Prior to this commit, we did not use the "is-k8s-installed"-role
everywhere where we wanted to know if Kubernetes is already set up on a
node. Instead, we used duplicated tasks in the roles which checked if the
"/etc/kubernetes" directory exists (not if it contains any useful files
or so, just if it exists).

This commit tries to figure out with a few more tests (and therefore
hopefully a higher chance of finding out the truth) if Kubernetes is
installed on the node (and the node joined the cluster). Additionally,
the "is-k8s-installed"-role is added as a dependency to roles which need
to know if a node has Kubernetes installed to remove the duplicated tasks.

Closes: #457