Allow to specify vars for custom stage in `config.toml`
These variables can be used in plays in the custom stage.

Copy the cluster-repo-specific inventory file from stage5.
This e.g. allows us referencing to the orchestrator host group
in custom plays.

Each host of the custom stage has access to the custom
vars specified in the dedicated section of the `config.toml`.

Closes: #462