Add skip for pip package installation
The `Ensure that kubernetes libraries are installed` tasks fails
during bare metal deployment on Ubuntu with following message:
`ERROR: Cannot uninstall 'PyYAML'. It is a distutils installed project
and thus we cannot accurately determine which files belong to it which
would lead to only a partial uninstall.`

As Ansible is executed within a venv and all PIP package installations
within our plays are installed within this venv, and the MetalController
does install the same packages, in order to execute our plays, we can
skip the installation if we detect that Ansible is being executed within
a venv.