ci: Spawn nodes with Ubuntu 18.04 and 20.04
Although we changed the default Ubuntu version to 20.04 we're still
officially supporting Ubuntu 18.04. I *think* that 20.04 starts docker
by default with the systemd-cgroup driver which is why we haven't
observed the failure reported in #457 before. This commit adds a bit of
variety to the CI to pick up problems more reliably.