k8s-tools: Remove system flag from pip install
It is not needed anymore and even breaks support with recent Debian
based distros like e.g. Ubuntu 2204.

The --system flag was nominally intended to allow installation in the
system package directory.

The "--system" flag is not needed for Debian-derived distros which
default to Python3 which should be the case for all OS in use now.