CI: OS project cleanup: Make subnet retrieval resilient
This reverts merge request !534.

In !534 and !539 we saw differing keys when trying to retrieve the list
of subnets from a network. We couldn't figure out why that was the case
because:

The hypothesis that the Neutron API changed can not be true. But also
the openstack client has not been updated or changed recently [1]. We
do not install the neutronclient and Terraform has not changed either.

It is unclear to me why we temporarily got the "subnet" key for the
network object instead of getting "subnet_ids".

[1] https://gitlab.com/yaook/k8s/-/merge_requests/539#note_959434771