Make centos networking settings broader available
`calico-node` pods panicking under certain conditions on Ubuntu, as they
are failing to obtain ip6tables rules.
From observations this is due to missing kernel modules.

Refactored and renamed the centos-networking role a bit, so it does also
apply to other distros.

Added also `net.bridge.bridge-nf-call-arptables` to the list of sysctl
entries, as it is actually part of the Kubernetes installation docs.

Adjusted the "Install the legacy network scripts" task so it is only
deployed on RHEL-like OSes.