Enable IOMMU in GRUB
To support VGPU passthrough to VMs, the IOMMU option on the worker
node has to be enabled.
A reboot block is needed if we use ansible via cloud-init otherwise
the node refuses to reboot himself.
Also both AMD and Intel CPUs are supported and the lcm sets the right
option depending on the CPU in the server.