CI: Fix before_script OpenStack project cleanup
Okay, here is a strange one. The before script OpenStack project cleanup
is broken, but only if it is executed from within docker or at least
it works when triggered locally although my local environment and the
docker one are the same. The Python versions are the same and also the
openstack client version is identical. I thought it may be because we
are not using a virtual environment in the docker container and
therefore created this merge request in k8s-ci [1], but that did not
help in this case.

The "delete_port()" function takes the name or id of the port and
returns a Boolean if the deletion was successful or not. It raises an
exception if the deletion fails.

We are passing the port as an object. This was no problem until now and
still works outside of a docker container. However, when debugging it
from within the docker container, I saw Neutron errors stating the the
port could not be found when the delete_port() function is triggered.
Explicitly passing the port.id fixes this issue.

[1] https://gitlab.com/yaook/images/k8s-ci/-/merge_requests/20